interface/requirements.txt
pandas
python-dotenv
pymongo
streamlit
streamlit-authenticator

